requirements.txt
python-dotenv
geopandas
eo-learn
sentinelsat
pyrootutils
hydra-core
typer
torchmetrics
albumentations
wandb
segmentation_models_pytorch
torchgeometry